refactor: replace component tests in cypress folder
